refactor(web): useAuth naming
